Update State to contain above and below sub-states
