+++ packages/editor/src/index.ts
@@ -5,5 +5,4 @@ import * as Util from "./util";
 export {reducer, print, Util};
 export * from "./editor-ast";
 export * from "./editor-reducer";
-export {isEqual} from "./util"; // TODO: dedupe methods in editor and util
-export {State} from "./state";
+export {isEqual, layoutCursorFromState} from "./util"; // TODO: dedupe methods in editor and util
