--- packages/editor/src/state.ts
@@ -1,14 +1,0 @@
-import * as Editor from "./editor-ast";
-
-import {LayoutCursor} from "./util";
-
-type ID = {
-    id: number;
-};
-
-export type State = {
-    math: Editor.Row<Editor.Glyph, ID>;
-    cursor: Editor.Cursor;
-    selectionStart?: Editor.Cursor;
-    cancelRegions?: LayoutCursor[];
-};
