+++ packages/editor/src/util.ts
@@ -1,5 +1,5 @@
 import * as Editor from "./editor-ast";
-import {State} from "./state";
+import {State} from "./above-reducer";
 
 export const isEqual = (
     a: Editor.Node<Editor.Glyph>,
